fix: 로그아웃 부분 수정
