+++ src/components/SubLayout.jsx
@@ -1,16 +1,6 @@
-import { Outlet, useNavigate } from "react-router-dom";
-import useAuthStore from "../zustand/authStore";
+import { Outlet } from "react-router-dom";
 
 const SubLayout = () => {
-  const navigate = useNavigate();
-
-  const { clearAuth } = useAuthStore();
-
-  const handleLogout = () => {
-    clearAuth();
-    navigate("/");
-  };
-
   return (
     <div className="flex flex-row w-full">
       <main className="flex-1">
